Create Descriptive Narrative

Create Descriptive Narrative

Usage

narrate_descriptive(
  df,
  measure = NULL,
  dimensions = NULL,
  coverage = 0.5,
  coverage_limit = 5,
  template_total = "{measure} across all {pluralize(dimension1)} is {total}. ",
  template_outlier = "Outlying {dimension} by {measure} is {outlier_insight}. ",
  template_outlier_multiple =
    "Outlying {pluralize(dimension)} by {measure} are {outlier_insight}. ",
  use_renviron = FALSE,
  return_data = FALSE,
  ...
)

Arguments

df

Data frame of tibble, can be aggregated or raw

measure

Numeric measure for function to create calculations with, if NULL then it will take the first numeric field available

dimensions

Vector of dimensions for analysis, by default all character or factor variable will be used

coverage

Portion of variability to be covered by narrative, 0 to 1

coverage_limit

Maximum number of elements to be narrated, overrides coverage to avoid extremely verbose narrative creation

template_total

glue template for total volumes narrative

template_outlier

glue template for single outlier narrative

template_outlier_multiple

glue template for multiple outliers narrative

use_renviron

Whether to use .Renviron variables in the template

return_data

return a list of variables used in the function's templates

...

other arguments passed to glue::glue

Value

character vector, glue

Examples

sales %>%
narrate_descriptive(measure = "Sales",
            dimensions = c("Territory", "Product"))
#> Sales across all Territories is 2 M. 
#> Outlying Territory by Sales is NA (943,899.4, 46.7 %). 
#>                             Outlying Products by Sales are Product E (465,134.6, 23 %), Product B (379,591.4, 18.8 %).