Create Descriptive Text Narratives from Data

narrate_descriptive() creates text narratives from a data frame containing one numeric and one or more character or factor text columns using glue syntax. Function can work with raw or aggregated data frame. It will automatically use the first numeric column as measure and all character or factor columns as dimensions

Usage

narrate_descriptive(
  df,
  measure = NULL,
  dimensions = NULL,
  summarization = "sum",
  coverage = 0.5,
  coverage_limit = 5,
  narration_depth = 2,
  template_total = "Total {measure} across all {pluralize(dimension_one)} is {total}.",
  template_average =
    "Average {measure} across all {pluralize(dimension_one)} is {total}.",
  template_outlier = "Outlying {dimension} by {measure} is {outlier_insight}.",
  template_outlier_multiple =
    "Outlying {pluralize(dimension)} by {measure} are {outlier_insight}.",
  template_outlier_l2 =
    "In {level_l1}, significant {level_l2} by {measure} is {outlier_insight}.",
  template_outlier_l2_multiple =
    "In {level_l1}, significant {pluralize(level_l2)} by {measure} are {outlier_insight}.",
  use_renviron = FALSE,
  return_data = FALSE,
  simplify = FALSE,
  format_numbers = TRUE,
  collapse_sep = ", ",
  collapse_last = " and ",
  ...
)

Arguments

df

data.frame() or tibble() Data frame of tibble, can be aggregated or raw

measure

Numeric measure for function to create calculations with, if NULL then it will take the first numeric field available

dimensions

Vector of dimensions for analysis, by default all character or factor variable will be used

summarization

Approach for data summarization/aggregation - 'sum', 'count' or 'average'

coverage

Numeric portion of variability to be covered by narrative, 0 to 1

coverage_limit

Integer maximum number of elements to be narrated, overrides coverage to avoid extremely verbose narrative creation

narration_depth

Parameter to control the depth of the analysis 1 for summary and 2 for detailed

template_total

glue template for total volumes narrative

template_average

glue template for average volumes narrative

template_outlier

glue template for single outlier narrative

template_outlier_multiple

glue template for multiple outliers narrative

template_outlier_l2

glue template for deeper hierarchical single outlier narrative

template_outlier_l2_multiple

glue template for deeper hierarchical multiple outliers narrative

use_renviron

If TRUE use .Renviron variables in the template. You can also set options(narrator.use_renviron = TRUE) to make it global for the session, or create an environment variable "use_renviron" by changing your .Renviron file usethis::edit_r_environ()

return_data

If TRUE - return a list of variables used in the function's templates

simplify

If TRUE - return a character vector, if FALSE - named list

format_numbers

If TRUE - format big numbers to K/M/B using format_num() function

collapse_sep

Separator for glue_collapse in cases with multiple values in single variable

collapse_last

Separator for glue_collapse for the last item, in cases with multiple values in single variable

...

other arguments passed to glue

Value

A list() of narratives by default and character() if simplify = TRUE

Examples

sales %>%
narrate_descriptive(measure = "Sales",
            dimensions = c("Region", "Product"))
#> $`Total Sales`
#> Total Sales across all Regions is 38.79 M.
#> 
#> $`Region by Sales`
#> Outlying Regions by Sales are NA (18.1 M, 46.6 %), EMEA (13.6 M, 34.9 %).
#> 
#> $`NA by Product`
#> In NA, significant Products by Sales are Food & Beverage (7.4 M, 40.9 %), Electronics (3.8 M, 21 %).
#> 
#> $`EMEA by Product`
#> In EMEA, significant Products by Sales are Food & Beverage (5.3 M, 38.8 %), Electronics (3.2 M, 23.5 %).
#> 
#> $`Product by Sales`
#> Outlying Products by Sales are Food & Beverage (15.5 M, 40.1 %), Electronics (8.6 M, 22.2 %).
#> 

sales %>%
  dplyr::filter(Product %in% c("Product A", "Product B")) %>%
  dplyr::group_by(Product, Region)  %>%
  dplyr::summarise(Quantity = sum(Quantity)) %>%
  narrate_descriptive()
#> $`Total Quantity`
#> Total Quantity across all Products is 0.
#> 
#> $`Product by Quantity`
#> Outlying Product by Quantity is .
#> 
#> $`Region by Quantity`
#> Outlying Region by Quantity is .
#> 

sales %>%
narrate_descriptive(measure = "Order ID", dimensions = "Region", summarization = "count")
#> $`Total Order ID`
#> Total Order ID across all Regions is 10 K.
#> 
#> $`Region by Order ID`
#> Outlying Regions by Order ID are NA (3,975, 39.8 %), EMEA (2,986, 29.9 %).
#>